master-node/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
click-spinner==0.1.10
croniter==1.4.1
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.2
fastapi-utilities==0.2.0
fastapi_amis_admin==0.7.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.3
passlib==1.7.4
pydantic==1.10.15
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlalchemy_database==0.1.2
sqlmodel==0.0.18
starlette==0.37.2
typer==0.12.3
typing_extensions==4.11.0
tzlocal==5.2
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
